+++ src/main/kotlin/offkilter/infohud/client/InfoHUDClient.kt
@@ -52,6 +52,10 @@ class InfoHUDClient : ClientModInitializer,InfoHUDSettings.Listener {
                     showHUD = !showHUD
                 }
             }
+
+            client.player?.let {
+                PerfCounters.updateSpeed(it)
+            }
         })
 
         InfoHUDSettings.INSTANCE.addListener(this)