+++ src/main/kotlin/offkilter/infohud/infoline/InfoLineRegistry.kt
@@ -31,4 +31,5 @@ object InfoLineRegistry {
     val ELYTRA = register(ElytraInfoLine())
     val ARMOR = register(ArmorInfoLine())
     val GAME_CLOCK = register(GameClockInfoLine())
+    val SPEED = register(SpeedInfoLine())
 }